requirements.txt
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
djoser==2.1.0
Pillow==9.1.1
stripe==3.4.0
gunicorn==20.1.0
whitenoise==6.2.0